% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NullC.R
\name{NullC}
\alias{NullC}
\title{Given a matrix M, find a matrix N giving a basis for the null space. This is a modified version of Null from the package MASS.}
\usage{
NullC(M)
}
\arguments{
\item{M}{An semi-orthogonal matrix}
}
\value{
A semi-orthogonal matrix N such that t(N) %*% M is a matrix of zeros.
}
\description{
Given a matrix M, find a matrix N giving a basis for the null space. This is a modified version of Null from the package MASS.
}
\note{
The MASS function Null(matrix(0,4,2)) returns a 4*2 matrix, whereas NullC(matrix(0,4,2)) returns diag(4).
}
\examples{
NullC(matrix(0,4,2))
}
