% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.stiefel.R
\name{optStiefel}
\alias{optStiefel}
\title{Optimize a function on the Stiefel manifold}
\usage{
optStiefel(F, dF, Vinit, method = "bb", searchParams = NULL,
  reltol = sqrt(.Machine$double.eps), maxIters = 100, verbose = FALSE,
  maxLineSearchIters = 100)
}
\arguments{
\item{F}{A function V(P, S) -> \code{R^1}}

\item{dF}{A function to compute the gradient of F.  Returns a \code{P \times S} matrix with \code{dF(X)_ij  = d(F(X))/dX_ij}}

\item{Vinit}{The starting point on the stiefel manifold for the optimization}

\item{method:}{"bb" or curvilinear}
}
\value{
A stationary point of F on the Stiefel manifold.
}
\examples{
## Find the first eigenspace spanned by the first P eigenvectors for a large matrix M
library(rstiefel)

N <- 1000
P <- 3
Lam <- diag(c(10, 5, 3, rep(1, N-P)))
U <- rustiefel(N, N)
M <- U \%*\% Lam \%*\% t(U)

F <- function(V) { - sum(diag(t(V) \%*\% M \%*\% V)) }
dF <- function(V) { - 2*M \%*\% V }
V = optStiefel(F, dF, Vinit=rustiefel(N, P),
               method="curvilinear",
               searchParams=list(rho1=0.1, rho2=0.9, tau=1))
               
print(sprintf("Sum of first \%d eigenvalues is \%f", P, -F(V)))

}
\references{
(Wen and Yin, 2013)
}
