% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbing.R
\name{rbing.matrix.gibbs}
\alias{rbing.matrix.gibbs}
\title{Gibbs Sampling for the Matrix-variate Bingham Distribution}
\usage{
rbing.matrix.gibbs(A, B, X)
}
\arguments{
\item{A}{A symmetric matrix}

\item{B}{A diagonal matrix with decreasing entries}

\item{X}{The current value of the random orthonormal Matrix}
}
\value{
A new value the matrix X obtained by Gibbs sampling
}
\description{
Simulate a random orthonormal matrix from the Bingham distribution using Gibbs sampling.
}
