% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbing.R
\name{rbing.O2}
\alias{rbing.O2}
\title{Efficiently simulate a 2x2 Orthogonal Random Matrix}
\usage{
rbing.O2(A, B, a = NULL, E = NULL)
}
\arguments{
\item{A}{A symmetric matrix.}

\item{B}{A diagonal matrix with decreasing entries.}

\item{a}{Sum of the eigenvalues of A, multiplied by the difference in B-values}

\item{E}{Eigenvectors of A.}
}
\value{
A random 2x2 orthogonal matrix simulated from the Bingham distribution.
}
\description{
Simulate a 2 x 2 random orthogonal matrix from the Bingham distribution using a rejection sampler.
}
