% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbing.R
\name{rbing.vector.gibbs}
\alias{rbing.vector.gibbs}
\title{Gibbs Sampling for the Vector-variate Bingham Distribution}
\usage{
rbing.vector.gibbs(A, x)
}
\arguments{
\item{A}{A symmetric matrix.}

\item{x}{The current value of the normal random vector.}
}
\value{
A new value of the vector \code{x} obtained by Gibbs sampling.
}
\description{
Simulate a random normal vector from the Bingham distribution using Gibbs sampling.
}
\note{
This provides one Gibbs scan. The function should be used iteratively.
}
