% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbmf.R
\name{rbmf.matrix.gibbs}
\alias{rbmf.matrix.gibbs}
\title{Gibbs Sampling for the Matrix-variate Bingham-von Mises-Fisher Distribution.
Simulate from the matrix bmf distribution as described in Hoff(2009).}
\usage{
rbmf.matrix.gibbs(A, B, C, X)
}
\arguments{
\item{A}{A symmetric matrix.}

\item{B}{A diagonal matrix with decreasing entries.}

\item{C}{A matrix with the same dimension as X.}

\item{X}{The current value of the random orthonormal matrix.}
}
\value{
A new value of the matrix \code{X} obtained by Gibbs sampling.
}
\description{
Gibbs Sampling for the Matrix-variate Bingham-von Mises-Fisher Distribution.
Simulate from the matrix bmf distribution as described in Hoff(2009).
}
\note{
This provides one Gibbs scan. The function should be used iteratively.    
WARNING: - do not start X at the eigenvectors of A: The relative weights on A then can become infinity.
}
\examples{
N <- 5; P <- 2;
X <- rustiefel(N, P)
U <- rustiefel(N, N)
A <- U \%*\% t(U)
B <- diag(N:1)
C <- rustiefel(N, P)
for(i in 1:100) {
 X <- rbmf.matrix.gibbs(A, B, C, X)
}
}
