% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbmf.R
\name{rbmf.vector.gibbs}
\alias{rbmf.vector.gibbs}
\title{Gibbs Sampling for the Vector-variate Bingham-von Mises-Fisher Distribution}
\usage{
rbmf.vector.gibbs(A, c, x)
}
\arguments{
\item{A}{A symmetric matrix}

\item{c}{A vector with the same length as \code{x}}

\item{x}{The current value of the random normal vector.}
}
\value{
A new value of the vector \code{x} obtained by Gibbs sampling.
}
\description{
Simulate a random normal vector from the Bingham-von Mises-Fisher distribution using Gibbs sampling as described in Hoff(2009) as described in Hoff(2009) as described in Hoff(2009) .
}
\note{
This is one Gibbs step, and must be used iteratively.
}
\examples{
N <- 5;
x <- rustiefel(N, 1)
U <- rustiefel(N, N)
A <- U \%*\% t(U)
c <- rustiefel(N, 1)
for(i in 1:100) {
   x <- rbmf.vector.gibbs(A, c, x)
}
}
