% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf.R
\name{rmf.matrix.gibbs}
\alias{rmf.matrix.gibbs}
\title{Gibbs Sampling for the Matrix-variate Bingham-von Mises-Fisher Distribution.}
\usage{
rmf.matrix.gibbs(M, X, rscol = NULL)
}
\arguments{
\item{M}{A matrix}

\item{X}{The current value of the random orthonormal matrix.}

\item{rscol}{The number of columns to update simultaneously.}
}
\value{
A new value of the matrix \code{X} obtained by Gibbs sampling.
}
\description{
Simulate a random orthonormal matrix from the matrix von Mises-Fisher distribution using Gibbs sampling, Hoff(2009).
}
\note{
This provides one Gibbs scan. The function should be used iteratively. The number of columns to replace should be small enough so that 
rmf.matrix will be quick, but not so small that you are not taking 
advantage of rmf.matrix. In particular, for square matrices you must 
be replacing two or more columns at a time, or else the chain is 
not irreducible.
}
\examples{
Z <- matrix(rnorm(10*5), 10, 5) 

U <- rmf.matrix(Z)
U <- rmf.matrix.gibbs(Z, U)
}
