% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rustiefel.R
\name{rustiefel}
\alias{rustiefel}
\title{Simulate a Uniformly Distributed Random Orthonormal Matrix

Simulate uniformly from \deqn{V(R, m)}
See Chikuse(1994)}
\usage{
rustiefel(m, R)
}
\arguments{
\item{m}{The length of each column vector.}

\item{R}{The number of column vectors.}
}
\value{
An \code{m x R}  uniformly distributed random semi-orthogonal matrix
}
