% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf.R
\name{rW}
\alias{rW}
\title{Simulate \code{W} as Described in Wood(1994)}
\usage{
rW(kap, m)
}
\arguments{
\item{kap}{A positive scalar.}

\item{m}{A positive integer.}
}
\value{
A number between zero and one.
}
\description{
Auxilliary variable simulation for rejection sampling of \code{rmf.vector}, 
as described in Wood(1994).
}
\examples{
rW(pi, 4)
}
