% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbing.R
\name{ry_bing}
\alias{ry_bing}
\title{Helper Function for Sampling a Bingham-distributed Vector}
\usage{
ry_bing(y, l)
}
\arguments{
\item{y}{A normal vector}

\item{l}{A vector}
}
\value{
A normal vector
}
\description{
C interface to perform a Gibbs update of 
\code{y} with invariant distribution  proportional to 
\deqn{\exp(sum(l*y^2))} with respect to the uniform measure on the sphere.
}
