% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbmf.R
\name{ry_bmf}
\alias{ry_bmf}
\title{Helper Function for Sampling a Bingham-von Mises-Fisher-distributed Vector}
\usage{
ry_bmf(y, l, d)
}
\arguments{
\item{y}{A normal vector.}

\item{l}{A vector}

\item{d}{A vector}
}
\value{
A normal vector
}
\description{
Helper Function for Sampling a Bingham-von Mises-Fisher-distributed Vector
}
