% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.stiefel.R
\name{lineSearch}
\alias{lineSearch}
\title{A curvilinear search on the Stiefel manifold (Wen and Yin 2013, Algo 1)}
\usage{
lineSearch(F, X, G_x, rho1, rho2, tauStart, maxIters = 100)
}
\arguments{
\item{F}{A function V(n, p) -> \code{R^1}}

\item{X}{an n x p semi-orthogonal matrix (starting point)}

\item{G_x}{an n x p matrix with \code{(G_x)_ij = dF(X)/dX_ij}}
}
\value{
A semi-orthogonal matrix, Ytau, which satisfies Armijo-Wolfe conditions
}
\description{
A curvilinear search on the Stiefel manifold (Wen and Yin 2013, Algo 1)
}
\references{
(Wen and Yin, 2013)
}
