% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.stiefel.R
\name{lineSearchBB}
\alias{lineSearchBB}
\title{A curvilinear search on the Stiefel manifold with BB steps (Wen and Yin 2013, Algo 2)
This is based on the line search algorithm described in (Zhang and Hager, 2004)}
\usage{
lineSearchBB(F, X, Xprev, G_x, G_xprev, rho, C, maxIters = 100)
}
\arguments{
\item{F}{A function V(n, p) -> R}

\item{X}{an n x p semi-orthogonal matrix}

\item{G_x}{an n x p matrix with (G_x)_ij = dF(X)/dX_ij}
}
\value{
A semi-orthogonal matrix Ytau which satisfies convergence criter (Eqn 29 in Wen & Yin '13)
}
\description{
A curvilinear search on the Stiefel manifold with BB steps (Wen and Yin 2013, Algo 2)
This is based on the line search algorithm described in (Zhang and Hager, 2004)
}
\references{
(Wen and Yin, 2013) and (Zhang and Hager, 2004)
}
