% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbmf.R
\name{rbmf.O2}
\alias{rbmf.O2}
\title{Simulate a \code{2 x 2} Orthogonal Random Matrix}
\usage{
rbmf.O2(A, B, C, env = FALSE)
}
\arguments{
\item{A}{A symmetric matrix.}

\item{B}{A diagonal matrix with decreasing entries.}

\item{C}{A \code{2 x 2} matrix.}

\item{env}{}
}
\value{
A random \code{2 x 2} orthogonal matrix simulated from the Bingham-von Mises-Fisher distribution.
}
\description{
Simulate a \code{2 x 2} Orthogonal Random Matrix
}
