% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rW.R
\name{rW}
\alias{rW}
\title{Simulate \code{W} as Described in Wood(1994)}
\usage{
rW(kap, m)
}
\arguments{
\item{kap}{a positive scalar.}

\item{m}{a positive integer.}
}
\value{
a number between zero and one.
}
\description{
Auxilliary variable simulation for rejection sampling of \code{rmf.vector},
as described in Wood(1994).
}
\examples{

rW(pi,4)

## The function is currently defined as
function (kap, m) 
{
    .C("rW", kap = as.double(kap), m = as.integer(m), w = double(1))$w
  }

}
\author{
Peter Hoff
}
