% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ry_bing.R
\name{ry_bing}
\alias{ry_bing}
\title{Helper Function for Sampling a Bingham-distributed Vector}
\usage{
ry_bing(y, l)
}
\arguments{
\item{y}{a normal vector.}

\item{l}{a vector.}
}
\value{
a normal vector.
}
\description{
C interface to perform a Gibbs update of \code{y} with invariant
distribution proportional to \code{exp( sum(l*y^2) } with respect to the
uniform measure on the sphere.
}
\examples{

## The function is currently defined as
function (y, l) 
{
    .C("ry_bing", y = as.double(y), l = as.double(l), n = as.integer(length(y)))$y
  }

}
\references{
Hoff(2009)
}
\author{
Peter Hoff
}
